Add partitioned_csv setup code to sql_integration test
