+++ datafusion/tests/sql/mod.rs
@@ -98,6 +98,7 @@ pub mod window;
 
 mod explain;
 pub mod information_schema;
+mod partitioned_csv;
 #[cfg_attr(not(feature = "unicode_expressions"), ignore)]
 pub mod unicode;
